refactor: move `diagnostics.Redact` to `redact.RedactSecrets`
